+++ src/generic/msgdlgg.cpp
@@ -73,7 +73,7 @@ wxGenericMessageDialog::wxGenericMessageDialog( wxWindow *parent,
     if (style & wxICON_MASK)
     {
          wxStaticBitmap *icon = new wxStaticBitmap(
-            this, -1, wxTheApp->GetStdIcon(style & wxICON_MASK));
+            this, -1, wxTheApp->GetStdIcon((int)(style & wxICON_MASK)));
          icon_text->Add( icon, 0, wxCENTER );
     }
